Configure remote script location and auto-remove

+++ lib/Benchmark/Remote/Payload.php
@@ -13,7 +13,9 @@
 namespace PhpBench\Benchmark\Remote;
 
 use PhpBench\Benchmark\Remote\Exception\ScriptErrorException;
+use RuntimeException;
 use Symfony\Component\Process\Process;
+
 
 /**
  * Class representing the context from which a script can be generated and executed by a PHP binary.
@@ -73,15 +75,29 @@ class Payload
     private $timeout;
 
     /**
+     * @var string
+     */
+    private $scriptPath;
+
+    /**
+     * @var bool
+     */
+    private $scriptRemove;
+
+    /**
      * Create a new Payload object with the given script template.
      * The template must be the path to a script template.
+     *
+     * @param array<string, mixed> $tokens
      */
     public function __construct(
         string $template,
         array $tokens = [],
-        $phpPath = PHP_BINARY,
+        ?string $phpPath = PHP_BINARY,
         ?float $timeout = null,
-        ProcessFactory $processFactory = null
+        ProcessFactory $processFactory = null,
+        string $scriptPath = null,
+        bool $scriptRemove = false
     ) {
         $this->setPhpPath($phpPath);
         $this->template = $template;
@@ -89,6 +105,9 @@ class Payload
         $this->processFactory = $processFactory ?: new ProcessFactory();
         $this->iniStringBuilder = new IniStringBuilder();
         $this->timeout = $timeout;
+        $this->phpPath = $phpPath;
+        $this->scriptPath = $scriptPath;
+        $this->scriptRemove = $scriptRemove;
     }
 
     public function setWrapper($wrapper): void
@@ -175,7 +194,30 @@ class Payload
 
     private function writeTempFile(string $script): string
     {
-        $scriptPath = tempnam(sys_get_temp_dir(), 'PhpBench');
+        $scriptPath = $this->scriptPath ?
+            $this->scriptPath . '/' . basename($this->template) :
+            tempnam(sys_get_temp_dir(), 'PhpBench');
+
+        if (false === $scriptPath) {
+            throw new RuntimeException(
+                'Could not generate temporary script name'
+            );
+        }
+
+        (function (string $directory): void {
+            if (file_exists($directory)) {
+                return;
+            }
+
+            if (@mkdir($directory, 0744)) {
+                return;
+            }
+
+            throw new RuntimeException(sprintf(
+                'Could not create directory "%s"', $directory
+            ));
+        })(dirname($scriptPath));
+
         file_put_contents($scriptPath, $script);
 
         return $scriptPath;
@@ -203,9 +245,16 @@ class Payload
 
     private function removeTmpFile(string $scriptPath): void
     {
+        if (!$this->scriptRemove) {
+            return;
+        }
+
         unlink($scriptPath);
     }
 
+    /**
+     * @return array<string, mixed>
+     */
     private function decodeResults(Process $process): array
     {
         $output = $process->getOutput();
